+++ components/dao_card2/index.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <n-link
-      :to="{name: 'voting-id', params: {id: card.pid}}"
+      :to="{name: 'daos-id', params: {id: card.pid}}"
       target="_blank"
     >
       <div class="dao-block">
