+++ pages/daos/_id.vue
